\name{pba-package}
\alias{pba-package}
\title{Performs bias analysis with probabilistic methods...}
\description{Performs bias analysis with probabilistic methods}
\details{\tabular{ll}{
Package: \tab pba\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2011-03-22\cr
Title: \tab Probabilistic Bias Analysis\cr
Author@R: \tab person("Jeremy", "Hetzel", "Thoms",email = "jthetzel@gmail.com")\cr
Author: \tab Jeremy T. Hetzel <jthetzel@gmail.com>\cr
Maintainer: \tab Jeremy T. Hetzel \email{jthetzel@gmail.com}\cr
Depends: \tab R (>= 2.12.0), plyr, ggplot2\cr
Description: \tab The pba package performs probabilistic bias analyses, 
with methods are adapted from Lash TL, Fox MP, and Fink AK's Applying 
Quantitative Bias Analysis to Epidemiological Data (2010). Currently,
only methods for adjustment of misclassification bias are included.\cr
License: \tab GPL-3\cr
URL: \tab http://pba.r-forge.r-project.org/\cr
BugReports: \tab https://r-forge.r-project.org/tracker/?func=add&group_id=988&atid=3913\cr		
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
pba: Probabilistic Bias Analysis

The pba package performs probabilistic bias analyses, with methods are 
adapted from Lash TL, Fox MP, and Fink AK's Applying Quantitative Bias 
Analysis to Epidemiological Data (2010). Currently, only methods for 
adjustment of misclassification bias are included.}
\docType{package}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
