\name{pba-package}
\alias{pba-package}
\title{pba: Probabilistic Bias Analysis...}
\description{pba: Probabilistic Bias Analysis}
\details{The pba package performs probabilistic bias analyses, with methods are 
adapted from Lash TL, Fox MP, and Fink AK's Applying Quantitative Bias 
Analysis to Epidemiological Data (2010). Currently, only methods for 
adjustment of misclassification bias are included.}
\docType{package}
\author{Jeremy Thoms Hetzel \email{jthetzel@gmail.com}}
